Refactor argument matchers rewriting into separate class
