+++ src/main/java/org/openrewrite/java/testing/jmockit/JMockitExpectationsToMockito.java
@@ -15,20 +15,17 @@
  */
 package org.openrewrite.java.testing.jmockit;
 
-import java.util.*;
-
 import lombok.EqualsAndHashCode;
 import lombok.Value;
-import org.openrewrite.Cursor;
-import org.openrewrite.ExecutionContext;
-import org.openrewrite.Preconditions;
-import org.openrewrite.Recipe;
-import org.openrewrite.TreeVisitor;
+import org.openrewrite.*;
 import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.java.JavaParser;
 import org.openrewrite.java.JavaTemplate;
 import org.openrewrite.java.search.UsesType;
 import org.openrewrite.java.tree.*;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Value
 @EqualsAndHashCode(callSuper = false)
@@ -82,34 +79,14 @@ public class JMockitExpectationsToMockito extends Recipe {
             }
             StringBuilder templateBuilder = new StringBuilder("verify(#{any(" + fqn + ")}, times(#{any(int)})).#{}(");
             for (Expression argument : arguments) {
-                templateBuilder.append(rewriteMethodArgument(ctx, argument));
+                templateBuilder.append(argument);
                 templateBuilder.append(", ");
             }
             templateBuilder.delete(templateBuilder.length() - 2, templateBuilder.length());
             templateBuilder.append(");");
             return templateBuilder.toString();
         }
-        private static final Set<String> JMOCKIT_ARGUMENT_MATCHERS = new HashSet<>();
-        static {
-            JMOCKIT_ARGUMENT_MATCHERS.add("anyString");
-            JMOCKIT_ARGUMENT_MATCHERS.add("anyInt");
-            JMOCKIT_ARGUMENT_MATCHERS.add("anyLong");
-            JMOCKIT_ARGUMENT_MATCHERS.add("anyDouble");
-            JMOCKIT_ARGUMENT_MATCHERS.add("anyFloat");
-            JMOCKIT_ARGUMENT_MATCHERS.add("anyBoolean");
-            JMOCKIT_ARGUMENT_MATCHERS.add("anyByte");
-            JMOCKIT_ARGUMENT_MATCHERS.add("anyChar");
-            JMOCKIT_ARGUMENT_MATCHERS.add("anyShort");
-            JMOCKIT_ARGUMENT_MATCHERS.add("any");
-        }
-        private static final Map<String, String> MOCKITO_COLLECTION_MATCHERS = new HashMap<>();
-        static {
-            MOCKITO_COLLECTION_MATCHERS.put("java.util.List", "anyList");
-            MOCKITO_COLLECTION_MATCHERS.put("java.util.Set", "anySet");
-            MOCKITO_COLLECTION_MATCHERS.put("java.util.Collection", "anyCollection");
-            MOCKITO_COLLECTION_MATCHERS.put("java.util.Iterable", "anyIterable");
-            MOCKITO_COLLECTION_MATCHERS.put("java.util.Map", "anyMap");
-        }
+
 
         @Override
         public J.MethodDeclaration visitMethodDeclaration(J.MethodDeclaration methodDeclaration, ExecutionContext ctx) {
@@ -150,6 +127,10 @@ public class JMockitExpectationsToMockito extends Recipe {
                     JavaCoordinates coordinates = nc.getCoordinates().replace();
                     J.Block expectationsBlock = (J.Block) nc.getBody().getStatements().get(0);
                     List<Object> templateParams = new ArrayList<>();
+
+                    // first rewrite the argument matchers
+                    ArgumentMatchersRewriter amr = new ArgumentMatchersRewriter(this, expectationsBlock, ctx);
+                    expectationsBlock = amr.rewrite();
 
                     // iterate over the expectations statements and rebuild the method body
                     int mockitoStatementIndex = 0;
@@ -254,10 +235,7 @@ public class JMockitExpectationsToMockito extends Recipe {
                 methodName = "doNothing";
                 templateParams.set(0, select);
                 templateParams.add(invocation.getName().getSimpleName());
-                mockArguments = rewriteMethodArgumentMatchers(ctx, invocation.getArguments());
-            } else {
-                templateParams.set(0,
-                        invocation.withArguments(rewriteMethodArgumentMatchers(ctx, invocation.getArguments())));
+                mockArguments = invocation.getArguments();
             }
             maybeAddImport("org.mockito.Mockito", methodName);
 
@@ -294,83 +272,6 @@ public class JMockitExpectationsToMockito extends Recipe {
             }
 
             return newBody;
-        }
-
-        private List<Expression> rewriteMethodArgumentMatchers(ExecutionContext ctx, List<Expression> arguments) {
-            List<Expression> newArguments = new ArrayList<>(arguments.size());
-            for (Expression methodArgument : arguments) {
-                newArguments.add(rewriteMethodArgument(ctx, methodArgument));
-            }
-            return newArguments;
-        }
-
-        private Expression rewriteMethodArgument(ExecutionContext ctx, Expression methodArgument) {
-            if (!isArgumentMatcher(methodArgument)) {
-                return methodArgument;
-            }
-            String argumentMatcher, template;
-            List<Object> argumentTemplateParams = new ArrayList<>();
-            if (!(methodArgument instanceof J.TypeCast)) {
-                argumentMatcher = ((J.Identifier) methodArgument).getSimpleName();
-                template = argumentMatcher + "()";
-                return rewriteMethodArgument(ctx, argumentMatcher, template, methodArgument, argumentTemplateParams);
-            }
-            J.TypeCast tc = (J.TypeCast) methodArgument;
-            argumentMatcher = ((J.Identifier) tc.getExpression()).getSimpleName();
-            String className, fqn;
-            JavaType typeCastType = tc.getType();
-            if (typeCastType instanceof JavaType.Parameterized) {
-                // strip the raw type from the parameterized type
-                className = ((JavaType.Parameterized) typeCastType).getType().getClassName();
-                fqn = ((JavaType.Parameterized) typeCastType).getType().getFullyQualifiedName();
-            } else if (typeCastType instanceof JavaType.FullyQualified) {
-                className = ((JavaType.FullyQualified) typeCastType).getClassName();
-                fqn = ((JavaType.FullyQualified) typeCastType).getFullyQualifiedName();
-            } else {
-                throw new IllegalStateException("Unexpected J.TypeCast type: " + typeCastType);
-            }
-            if (MOCKITO_COLLECTION_MATCHERS.containsKey(fqn)) {
-                // mockito has specific argument matchers for collections
-                argumentMatcher = MOCKITO_COLLECTION_MATCHERS.get(fqn);
-                template = argumentMatcher + "()";
-            } else {
-                // rewrite parameter from ((<type>) any) to <type>.class
-                argumentTemplateParams.add(JavaTemplate.builder("#{}.class")
-                        .javaParser(JavaParser.fromJavaVersion())
-                        .build()
-                        .apply(
-                                new Cursor(getCursor(), tc),
-                                tc.getCoordinates().replace(),
-                                className
-                        ));
-                template = argumentMatcher + "(#{any(java.lang.Class)})";
-            }
-            return rewriteMethodArgument(ctx, argumentMatcher, template, methodArgument, argumentTemplateParams);
-        }
-
-        private Expression rewriteMethodArgument(ExecutionContext ctx, String argumentMatcher, String template,
-                                                 Expression methodArgument, List<Object> templateParams) {
-            maybeAddImport("org.mockito.Mockito", argumentMatcher);
-            return JavaTemplate.builder(template)
-                    .javaParser(JavaParser.fromJavaVersion().classpathFromResources(ctx, "mockito-core-3.12"))
-                    .staticImports("org.mockito.Mockito." + argumentMatcher)
-                    .build()
-                    .apply(
-                            new Cursor(getCursor(), methodArgument),
-                            methodArgument.getCoordinates().replace(),
-                            templateParams.toArray()
-                    );
-        }
-
-        private static boolean isArgumentMatcher(Expression expression) {
-            if (expression instanceof J.TypeCast) {
-                expression = ((J.TypeCast) expression).getExpression();
-            }
-            if (!(expression instanceof J.Identifier)) {
-                return false;
-            }
-            J.Identifier identifier = (J.Identifier) expression;
-            return JMOCKIT_ARGUMENT_MATCHERS.contains(identifier.getSimpleName());
         }
 
         private static String getMockitoStatementTemplate(Expression result, String fqn,
